Style Guide - Quotes
http://docs.puppetlabs.com/guides/style_guide.html#quoting
All strings that do not contain variables should be enclosed in single
quotes.  Double quotes should be used when variable interpolation is
required.  Quoting is optional when the string is an alphanumeric bare
word and is not a resource title.

All variables should be enclosed in braces when interpolated in a
string.