Support upstart in VMware Tools 5.1 on EL6
VMware Tools 5.1 on EL6 is now using upstart and not System V init.
Unfortunately, as of version 3.0.1, Puppet does not support Upstart on
any operatingsystem except Ubuntu.
http://projects.puppetlabs.com/issues/11989
http://projects.puppetlabs.com/issues/14036
This commit hacks around the lack of Puppet support for Upstart on
RedHat by providing custom start, stop, restart, and status commands to
the service resource.